Fix Metadata Light Up in Windows Store Apps, small API review issue, 
and compiler warning.

+++ src/System.Reflection.Metadata/src/System/Reflection/Metadata/Handles.cs
@@ -179,7 +179,7 @@ namespace System.Reflection.Metadata
             // All virtual tokens will be sorted after non-virtual tokens.
             // The order of handles that differ in kind is undefined, 
             // but we include it so that we ensure consistency with == and != operators.
-            return ((long)left._value | (long)left._vType << 32).CompareTo((long)right._value | (long)right._vType << 32);
+            return ((long)(uint)left._value | (long)left._vType << 32).CompareTo((long)(uint)right._value | (long)right._vType << 32);
         }
 
         [MethodImpl(MethodImplOptions.NoInlining)]
